% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bib.dropdown}
\alias{bib.dropdown}
\title{List selector options from bib file, in alphabetical order}
\usage{
bib.dropdown(
  fnm = "~/texmf/bibtex/bib/master-refs.bib",
  field = "keywords",
  add = NA,
  string.out = F
)
}
\arguments{
\item{fnm}{Filename of .bib file (including path). Default is ~/PhD/References/Master-refs.bib}

\item{field}{Field to interrogate: default is 'keywords'}
}
\value{
Vector of sorted keywords
}
\description{
List selector options from bib file, in alphabetical order
}
\details{
eg. to extract all keywords stored in the 'manager' menu for easier revision
}
\examples{
kw <- bib.dropdown("~/PhD/References/Master-refs.bib")
paste(kw, collapse = ";")

}
