% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{boot.sig}
\alias{boot.sig}
\title{Check where target value falls within bootstrap sample}
\usage{
boot.sig(boot.samp, target = 0, q1 = 0.025, q2 = 0.975)
}
\arguments{
\item{boot.samp}{Vector of bootstrapped statistic, as returned by \link{\code{bootstrap}}}

\item{target}{Target value. Default is 0.}

\item{q1}{Lower probability bound. Default is 0.025.}

\item{q2}{Upper probability bound. Default is 0.975.}
}
\value{
Value: either 0 (target in lower tail), 1 (target not significant), or 2 (target in upper tail)
}
\description{
Check where target value falls within bootstrap sample
}
