% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bootstrap}
\alias{bootstrap}
\title{Get bootstrap distribution of a statistic over a set of values}
\usage{
bootstrap(x, nsamp = 10000, fn = mean)
}
\arguments{
\item{x}{Vector or array of values}

\item{nsamp}{Number of bootstrap samples to produce. Default is 10000.}

\item{fn}{Statistic to calculate for each bootstrap sample. Default is mean.}
}
\value{
nsamp-length vector of bootstrapped statistics
}
\description{
Get bootstrap distribution of a statistic over a set of values
}
