% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gubbins.R
\name{cb.out}
\alias{cb.out}
\title{open image plot, ready to output from later commands}
\usage{
cb.out(fformat, filenm, mar = c(2, 2, 1, 1), ...)
}
\description{
open image plot, ready to output from later commands
}
\examples{
cb.out("jpeg", "./tmp")
plot(which(pw.m[,,"black", "loan"] > 10000, arr.ind = T))
dev.off()

}

