% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clean.tex}
\alias{clean.tex}
\title{Remove LaTeX production files (.aux, .log etc)}
\usage{
clean.tex(fpath)
}
\arguments{
\item{fpath}{File path to clear files from}
}
\description{
Remove LaTeX production files (.aux, .log etc)
}
