% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gubbins.R
\name{clump.centres}
\alias{clump.centres}
\title{Get centre & radius of clumps}
\usage{
clump.centres(px)
}
\arguments{
\item{px}{Matrix of coordinates}
}
\description{
Given a list of coordinates of points, identify clusters and for each, find the midpoint and spanning radius.
}

