% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cool.colors}
\alias{cool.colors}
\title{Colour palette of cool colours (designed as companion to heat.colors)}
\usage{
cool.colors(n, bias = 1, rev = F)
}
\description{
Colour palette of cool colours (designed as companion to heat.colors)
}
