% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{crop.pdf}
\alias{crop.pdf}
\title{Crop a pdf}
\usage{
crop.pdf(filenm)
}
\arguments{
\item{filenm}{String giving filename to be cropped (including full path and extension)}
}
\description{
Call 'pdfcrop' function without having to switch to terminal
}
