% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.boxplot}
\alias{.boxplot}
\title{Boxplot with less annoying defaults}
\usage{
.boxplot(x, cex = 0.4, main = "", xlab = "", ylab = "", hline = NA, ...)
}
\description{
Boxplot with less annoying defaults
}
