% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.matplot}
\alias{.matplot}
\title{Matplot with less annoying defaults}
\usage{
.matplot(x, y, xlab = "", ylab = "", type = "l", lty = 1, pch = 20, ...)
}
\description{
Matplot with less annoying defaults
}
