% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gubbins.R
\name{fancy.date}
\alias{fancy.date}
\title{convert integer date to string, with dashes}
\usage{
fancy.date(dt)
}
\arguments{
\item{dt}{Integer date, in format yymmdd}
}
\value{
String date, yy-mm-dd
}
\description{
convert integer date to string, with dashes
}

