% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{filled.band}
\alias{filled.band}
\title{Filled band between two lines}
\usage{
filled.band(x, y.max, y.min, col)
}
\arguments{
\item{x}{Vector of x values for plotting}

\item{y.max}{Vector of y values for upper line}

\item{col}{Colour for shading}

\item{y.mix}{Vector of y values for lower line}
}
\description{
Filled band between two lines
}
