% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{filled.ellipse}
\alias{filled.ellipse}
\title{Add shaded ellipse to plot}
\usage{
filled.ellipse(ell, ...)
}
\arguments{
\item{ell}{Ellipse to be drawn}

\item{...}{Additional graphical parameters to be passed to plotting function}
}
\description{
Add shaded ellipse to plot
}
