% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gamm.pars}
\alias{gamm.pars}
\title{Gamma parameters given specified mean/mode and variance}
\usage{
gamm.pars(mode, var, mean)
}
\arguments{
\item{mode}{Mode of Gamma distribution.}

\item{var}{Variance of Gamma distribution.}

\item{mean}{Mean of Gamma distribution. Only used if mode is not provided.}
}
\value{
List containing shape and rate parameters.
}
\description{
Gamma parameters given specified mean/mode and variance
}
