% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{greyscale}
\alias{greyscale}
\title{Convert a pdf to greyscale}
\usage{
greyscale(fnm.in, fnm.out = gsub("\\\\.pdf", "-GS.pdf", fnm.in),
  open = F)
}
\arguments{
\item{fnm.in}{Filename of pdf to convert to greyscale, including path}

\item{fnm.out}{Filename of converted pdf, including path. If not supplied, will simply append '-GS' to the input filename.}

\item{open}{Boolean: open the converted pdf? Default is N.}
}
\description{
Convert a pdf to greyscale
}
