% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gubbins.R
\name{image.scale}
\alias{image.scale}
\title{plot separate image scale}
\usage{
image.scale(z, zlim, col = heat.colors(12), breaks, axis.pos = 1,
  add.axis = TRUE, ...)
}
\description{
Plot rectangular image scale, based on given values & colours
}

