% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ix}
\alias{ix}
\title{Integer as 2-character string}
\usage{
ix(i, ...)
}
\arguments{
\item{i}{Integer to be converted to 2-character string}
}
\value{
String of length 2
}
\description{
Integer as 2-character string
}
