% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{list.repos}
\alias{list.repos}
\title{List all git repositories}
\usage{
list.repos(folders = c("~/R/My-packages/", "~/PhD/", "~/texmf/"), all = F)
}
\arguments{
\item{folders}{Filepath to begin searching for git repositories. Default is c("~/R/My-packages/", "~/PhD/").}
}
\value{
Vector of paths to git repositories
}
\description{
List all git repositories
}
\examples{
gl <- list.repos()

}
