% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{makepdf}
\alias{makepdf}
\title{Quickly produce a pdf}
\usage{
makepdf(fnm, plotdef, crop = T, margin = c(0, 0, 0, 0), ...)
}
\arguments{
\item{fnm}{String: path & name of pdf to be created}

\item{plotdef}{Define function to be plotted}

\item{crop}{Boolean: apply pdfcrop to the final image?}

\item{...}{Further arguments to be passed to pdf function}
}
\description{
Quickly produce a pdf
}
\examples{
makepdf("./Documents/tmp.pdf", width = 4, {
     plot(1:100, dnorm(1:100, 50,20))
})

}
