% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{matcheck}
\alias{matcheck}
\title{Check matrix equivalence}
\usage{
matcheck(expr1, expr2, dp = 12, seed = 1)
}
\arguments{
\item{expr1}{First matrix equation; either a string using A,B,C,D or a formula using matrices in the global environment}

\item{expr2}{Second matrix equation; either a string using A,B,C,D or a formula using matrices in the global environment}

\item{dp}{Number of significant figures to check result to; default is 12.}

\item{seed}{Integer: seed used to generate random matrices}
}
\description{
Confirm whether two matrix equations produce identical output
}
