% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hijacks.R
\name{panel.cor}
\alias{panel.cor}
\title{Calculate correlation for scatterplot matrix}
\usage{
panel.cor(x, y, digits = 2, cex.cor = 1, ...)
}
\arguments{
\item{x}{First variable to compare}

\item{y}{Second variable to compare}

\item{digits}{number of decimal places to display}

\item{cex.cor}{Scaling factor to apply to text}
}
\description{
Calculate per-variable correlation & p-value, to be used as panel input in \code{\link{pairs}} etc.
}
\examples{
pairs(iris[1:4], pch = 21, bg = c("red", "green3", "blue")[unclass(iris$Species)], upper.panel = panel.cor)

}
