% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gubbins.R
\name{prep.csv}
\alias{prep.csv}
\title{Replace 0 with "-"}
\usage{
prep.csv(qq, dp = 1)
}
\arguments{
\item{qq}{Object to be prepared}

\item{dp}{Decimal places desired}
}
\value{
Original object, with 0 (and, incidentally, NA) replaced with "-"
}
\description{
Prepare an object (usually a table or data frame) for output to csv for LaTeX, by rounding to set DP and replacing 0 with "-".
}

