% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{px.in.ellipse}
\alias{px.in.ellipse}
\title{Identify whether points fall within an ellipse}
\usage{
px.in.ellipse(px, mu, Sigma, p = 0.95)
}
\arguments{
\item{px}{Points to be tested}

\item{mu}{Centre of ellipse}

\item{Sigma}{Covariance matrix describing ellipse}

\item{p}{Define proportion of density to include}
}
\description{
Identify whether points fall within an ellipse
}
