% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hijacks.R
\name{r.invwishart}
\alias{r.invwishart}
\title{Randomly generate a single sample from an inverse-Wishart distribution (robust)}
\usage{
r.invwishart(nu, S)
}
\arguments{
\item{nu}{Scalar degrees of freedom}

\item{S}{Symmetric, positive-semidefinite scale matrix (may be rank deficient)}
}
\description{
Randomly generate a single sample from an inverse-Wishart distribution (robust)
}
