% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hijacks.R
\name{r.norminvwishart}
\alias{r.norminvwishart}
\title{Randomly generate a single sample from a normal-inverse-Wishart distribution (robust)}
\usage{
r.norminvwishart(mu0, lambda, S, nu)
}
\arguments{
\item{mu0}{Mean vector}

\item{lambda}{Positive scalar used to obtain V(mu) = Sigma/lambda}

\item{S}{Symmetric, positive-semidefinite scale matrix (may be rank deficient)}

\item{nu}{Scalar degrees of freedom}
}
\description{
Randomly generate a single sample from a normal-inverse-Wishart distribution (robust)
}
