% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ratty.weights}
\alias{ratty.weights}
\title{Update ratty weight graphs with latest data in Google Sheets}
\usage{
ratty.weights()
}
\description{
Update ratty weight graphs with latest data in Google Sheets
}
