% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{repo.status}
\alias{repo.status}
\title{Check status of git repository}
\usage{
repo.status(repo.nm)
}
\arguments{
\item{repo.nm}{File path to git repository}
}
\value{
Vector of status checks
}
\description{
Check status of git repository
}
\examples{
t(sapply(gl, repo.status))

}
