% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gubbins.R
\name{rmerge.df.list}
\alias{rmerge.df.list}
\title{Recursively merge list of data frames}
\usage{
rmerge.df.list(head, df.list, ...)
}
\arguments{
\item{head}{Data frame with which all others are to be merge (for example, a list of coordinates, to which classifications can be added)}

\item{df.list}{List of additional data frames to be merged with the head}

\item{...}{Additional optional parameters to be passed to the \link{\code{merge}} function.}
}
\value{
Data frame containing merged data
}
\description{
Recursively merge list of data frames with a 'master data frame'
}

