% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{source.lines}
\alias{source.lines}
\title{Source a block of lines from a named file}
\usage{
source.lines(fnm, from, to)
}
\arguments{
\item{fnm}{Filename to source}

\item{from}{Line to start sourcing from}

\item{to}{Line to stop sourcing}
}
\description{
Source a block of lines from a named file
}
