% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{temp.cols}
\alias{temp.cols}
\title{Quick palette of colours from cool to warm, separated by white}
\usage{
temp.cols(ncool, nwarm, mid.col = "white")
}
\arguments{
\item{ncool}{Number of colours to generate from cool.colors()}

\item{nwarm}{Number of colours to generate from heat.colors()}
}
\description{
Quick palette of colours from cool to warm, separated by white
}
