% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{test.colramp}
\alias{test.colramp}
\title{Draw test square showing colour ramp}
\usage{
test.colramp(mycolramp, n = 100)
}
\description{
Test function: plot a test square, shaded according to a given colour ramp.
}
