% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tidyup}
\alias{tidyup}
\title{Clear environment, except for selected objects}
\usage{
tidyup(keep = "")
}
\arguments{
\item{keep}{Vector of names of objects to keep}
}
\description{
Clear environment, except for selected objects
}
