% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{transp}
\alias{transp}
\title{Wrapper function to get vector of transparent colours}
\usage{
transp(cols, alpha = 0.4)
}
\arguments{
\item{cols}{Vector of colours to adjust}

\item{alpha}{Number between 0 and 1 indicating required transparency. Default is 0.4}
}
\value{
Vector of colours, adjusted to desired transparency.
}
\description{
Wrapper function to get vector of transparent colours
}
