% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unimode}
\alias{unimode}
\title{Return mode of x}
\usage{
unimode(x)
}
\arguments{
\item{x}{Vector or matrix of values - assumed to be unimodal}
}
\value{
Value of (single) mode of x
}
\description{
Return mode of x
}
