% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gubbins.R
\name{update.functions}
\alias{update.functions}
\title{Refresh package with latest changes}
\usage{
update.functions(package = "IO.pixels")
}
\arguments{
\item{package}{String: name of package to be re-documented and reinstalled.}
}
\description{
Re-document and reinstall a local package
}
\examples{
update.functions("IO.Pixels")


}

