% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hijacks.R
\name{whichmax}
\alias{whichmax}
\title{Identify maximum value}
\usage{
whichmax(x, na.rm = T, arr.ind = T, ...)
}
\arguments{
\item{x}{Vector or array of values to compare}

\item{...}{Additional arguments to pass to which()}
}
\description{
Replaces base function with possibility to return indices
}
