% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{list.repos}
\alias{list.repos}
\title{List all git repositories}
\usage{
list.repos(fpath = "/home/")
}
\arguments{
\item{fpath}{Filepath to begin searching for git repositories. Default is /home/.}
}
\value{
Vector of paths to git repositories
}
\description{
List all git repositories
}
